Implement `DynamicPanelController`
