+++ Assets/Project/Scripts/GamePlayScene/Panel/DynamicPanelController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 950d269fc9172489aa122df63e8d3008
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
