+++ Assets/Project/Scripts/GamePlayScene/Panel/PanelController.cs
@@ -4,13 +4,5 @@ namespace Project.Scripts.GamePlayScene.Panel
 {
 	public abstract class PanelController : MonoBehaviour
 	{
-		protected GamePlayDirector gamePlayDirector;
-
-		protected virtual void Start()
-		{
-			gamePlayDirector = FindObjectOfType<GamePlayDirector>();
-		}
-
-		public abstract void Initialize(GameObject finalTile);
 	}
 }
